+++ libft/values_abs.c
@@ -6,15 +6,15 @@
 /*   By: jjuntune <jjuntune@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/27 14:30:19 by jjuntune          #+#    #+#             */
-/*   Updated: 2022/04/27 14:36:45 by jjuntune         ###   ########.fr       */
+/*   Updated: 2022/04/27 18:48:44 by jjuntune         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	values_abs(int i)
+double	values_abs(double i)
 {
 	if (i < 0)
-		i *= -1;
+		i *= -1.0;
 	return (i);
 }
